requirements.txt
absl-py==1.4.0
cython==0.29.32
emukit==0.4.10
GPy==1.10.0
graphviz==0.20.1
matplotlib==3.7.1
numpy==1.23.1
networkx==3.1
paramz==0.9.5
pygraphviz==1.11
scipy==1.10.1
scikit-learn==1.2.2
typing_extensions==4.6.3
ml-collections==0.1.1
